May 11 2014 @12:44 AM

+++ Abacus/AbacusView.m
@@ -11,7 +11,6 @@
 #import "RowView.h"
 #import "Column.h"
 #import "UIConfig.h"
-#import "UIConstants.h"
 #import "AbacusConstants.h"
 
 @implementation AbacusView
@@ -48,7 +47,7 @@
 
 -(void)drawColumns:(CGRect)rect
 {
-    int placeValue = 1;
+    NSInteger placeValue = 1;
     for(int i = 1; i <= NUM_OF_COLUMNS; i++) {
         CGRect columnViewBounds = [UIConfig getColumnViewBounds:rect forColumn:i];
         ColumnView *columnView = [[ColumnView alloc] initWithFrame:columnViewBounds placeValue:placeValue];
